+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.4.0 (2016-02-16)
 
+Adds:
+
+  - Adds 'application' and 'local?' helper methods to PansophyAuthenticator module
+  
 Fixes:
 
   - Tests could be influenced by system environment variables